Fix _nanos6_pre_loader.
* loader/loader.c : * loader/loader.c : Add extra check for
argc > 1  argv[0] != NULL because it seems to be an
inconsistency with the two values.

Macro condition _nanos6_pre_loader to build only in cluster
compilation. As this function is called very early we can't condition
this code in runtime; but at least we are sure we only need it in
cluster mode.