Big lock around all calls to Extrae
There is a race condition that happens when Nanos6 creates a new thread.
Doing so eventually calls Backend_ChangeNumberOfThreads() in Extrae,
which reallocates various buffers without taking a lock. A concurrent
write to these buffers by a different thread often causes a segmentation
fault, for example in Buffer_InsertMultiple in Extrae. This change may
introduce a performance overhead.