CPUObjectCache returns surplus free objects
If CPUObjectCache::newObject and deleteObject happen on different CPUs
then surplus free objects on a CPU must be returned to the
NUMAObjectCache.  This happens commonly when tasks and associated data
structures are created in a parent task on one CPU and deleted on an
arbitrary CPU. This code was originally developed by Jimmy Aguilar.