Non-object memory allocators not in per-CPU pool
MemoryAllocator::alloc and free now have an extra parameter that
specifies whether to use the per-CPU pool or not (the default).  Only
use the per-CPU pool if (1) performance critical so important not to
take a lock, and (2) either alloc and free happen on the same CPU or
there is a mechanism (outside the MemoryAllocator) to redistribute free
memory among CPUs. Currently only the CPUObjectCache satisfies these
requirements and uses the per-CPU pool. Previously all allocations on a
worker thread used the per-CPU pool, leading to excessive memory use.