Fixed missing lock on external memory pool
The external pool is shared among cores, so MemoryAllocator::alloc
and free need to take the _externalMemoryPoolLock when they
call pool->getChunk() or pool->returnChunk() on it. This bug
was exposed in the previous commit (21b0f686), which greatly
increased the number of allocations and frees on the external pool.