Add nanos6_get_app_communicator()
This Nanos6 API function returns the MPI communicator to be used by the
application. The communicator is returned by reference via a parameter, in a
similar style to MPI. It is a void * parameter (rather than MPI_Comm *) so that
cluster.h, which declares the function, can be used by a program that doesn't
include <mpi.h>. This doesn't exactly match ICPP 2022, but is a better
solution.

Note we cannot declare the function conditionally depending on whether
MPI_VERSION or a similar preprocessor variable is defined, since cluster.h may
be included before mpi.h (in fact it will be, as Mercurium includes the Nanos6
API include files automatically, before the user's code). The function will
fail if it is called if the runtime isn't built with cluster (as we don't have
mpi.h to be able to return an MPI_Comm). In pure cluster mode, we return
MPI_COMM_SELF, since the program shouldn't do its own MPI communication. In
non-cluster mode, we return MPI_COMM_WORLD, since in that case the program can
use MPI however it likes.