ClusterBalanceScheduler: load balancing scheduler
This scheduler attempts to schedule tasks with load balancing among the nodes.
It assumes a relatively large number of "strong" tasks.

It provisionally schedules each task to a node using the locality scheduler. If
the best node has fewer than two offloaded tasks per core, then the task is
offloaded to it immediately.  Otherwise the least "busy" node is chosen,
according to the number of offloaded tasks. If the node has fewer than two
offloaded tasks per core, then the task is offloaded to it immediately.
Otherwise the task is queued inside the balance scheduler. It will be taken
from the queue (with work stealing) whenever a local or remote task finishes.

The balance scheduler can be selected using the config:
cluster.scheduling_policy = "balance"