Fix MessageSatisfiability when location is nullptr
There was an occasional segfault in ClusterDataLinkStep::linkRegion(), which
happened when the location to be sent in a MessageSatisfiability message was
nullptr (rather than pointing to a valid MemoryPlace). This happens very rarely
when write satisfiability is propagated before read satisfiability.

TaskOffloading::SatisfiabilityInfo::_src is -1 when the location is nullptr.