Dmalloc: distribute using file interface
ClusterNode holds the number of cores for the Nanos6 instance, and this is
taken into account when a Dmalloc array is distributed.  The .hybrid/apprank<n>
file (e.g. .hybrid/apprank for apprank 0) contains a list of the #cores for each
instance in the apprank. This can be on a single line, separated by spaces.

The global core allocation is used to (1) distribute dmalloc affinity and (2)
as one of the terms in how many tasks to request for remote work stealing.
Start with 1 for current node and 0 for other nodes => dmalloc starts not
distributed (from the point of view of each instance) and no tasks are
initially stealed from other instances.