nanos6_dmalloc: redistribute according to the number of cores
ClusterNode holds the number of cores for the Nanos6 instance, and this is
taken into account when a Dmalloc array is distributed.

Implement reading of the number of cores per instance for the global allocation
policy. The .hybrid/apprank<apprank> file contains a list giving the number of
cores for each instance in the apprank.