Support redistribution of dmallocs
Rredistribution of dmallocs is necessary for DLB integration (when the number
of cores per node changes) and for malleability (when the number of nodes
changes). The runtime needs to call ClusterMemoryManagement::
redistributeMallocs whenever redistribution is required. Each dmalloc is
recorded in a list, and on redistribution it will be unregistered then again
registered.  A write lock is taken on the directory to avoid race conditions
with other threads that may otherwise be reading (or writing!) the directory.
The list of valid dmallocs is also used to raise an error if a dfree is done on
an invalid region.