Hybrid+DLB: start with a sensible CPU mask
In MPI + OmpSs-2@Cluster mode, the initial CPU ownership allocation is as
follows: each helper process on the node gets one core and the remaining cores
are divided among the master processes. If there are no master processes, then
the cores are divided among the helper processes instead. Any leftover cores
(due to rounding) are given to the last external rank on the node, so that all
cores start with an owner.