Only strong accesses to unknown memory raise an error
Previously any access to an unknown region of memory used to fail with the
"unkown region" error. This change makes it so that a weak access to unknown
memory is always OK, but it is an error to have a strong access on unknown
memory.  In the old code, Directory::find directly raised the error as soon as
it saw a missing region.  This change makes Directory::find return an empty
list if the region isn't in the directory, so it is up to the caller to decide
what to do. The scheduler ignores unknown regions for scheduling purposes. A new
debug check (in the debug build only) checks whether all of the task's strong
accesses are in the directory, and it generates an error message if any of them
aren't.