Update Extrae for MPI + OmpSs-2@Cluster
Renumber the MPI ranks in the trace, so that they are ordered firstly by
apprank and then by internal rank (within the apprank). This order is easy to
interpret, unlike the default process ordering from MPI_COMM_WORLD. If running
pure OmpSs-2@Cluster, i.e. with one apprank, then the ordering is the same as
before. Reordering is needed when there is only one cluster node in the
instance, since there may be more than one apprank. The behaviour is only
disabled if cluster.communication is set to "disabled".