Add stats instrumentation of cluster messages
For stats instrumentation, add a breakdown of the total number of messages and
total number of bytes, for each type of message. This requires new
instrumentation calls for raw data messages, which are used for task data
transfers. Unfortunately raw data messages are also used to set up the virtual
memory, and this happens before the nstrumentation subsystem is initialized. To
avoid errors at initialization, an extra 'instrumentation' argument is added to
the fetchData and sendData methods.