Avoid redundant data fetches in cluster
Avoid sending multiple concurrent data fetches for the same region, for example
when multiple ready tasks are offloaded with the same in dependency. This is
done by checking the list of pending DataTransfers to see if there is already
an incoming DataTransfer for the same region or a region fully containing it.
Most of the functionality is inside the cluster execution workflow, which calls
a new checkPendingQueue method of ClusterPollingServices::PendingQueue.