Remove incorrect assert on read before write satisfiability
This assertion was checking that read satisfiability arrives
before write satisfiability, but there is in fact a benign race
condition, which means that it is indeed possible for write
satisfiability to arrive before read satisfiability. This assertion
was triggering in the "nasty" tests on a small number of nodes,
which have a lot of very small tasks.