Avoid redundant data fetches for in dependencies
The previous version performed a separate data fetch for every task in the
following example program (illustrative of Barnes--Hut):

// write big array on node(0)
for(...) {
  #pragma oss task inout(<small array>) in(<big array>) node(1)
  { ... }
}

This happened because the location is passed in the dependency system with read
satisfiability. In this example, tasks have read satisfiability when they are
created, so each offloaded task is told that the location is node 0, (a) there
is only one location in the DataAccess class (this is the easiest issue to fix
as in the clusters version we could of course extend the location to a set of
nodes), (b) read satisfiability is propagated before the task is even scheduled
so we don't yet know where the task will be executed, and (c) it is not safe to
put node 1 as an alternative location for the data until the data actually
arrives at node 1, which requires some sort of message from node 1. Then since
the tasks will have already been offloaded, it seems lots of new messages would
be needed, to then pass the new location through the dependency system (in this
specific case of multiple readers).

The basic idea is to choose a new 64-bit random number for every data access
that is not READ_ACCESS_TYPE (i.e. not an in dependency). An access may be
fragmented, but (write ID, region) is always unique, assuming no 64-bit write
ID collisions (see discussion in WriteID.hpp). This write ID is passed among
nodes, and each node is responsible for keeping track of which (write ID,
region) combinations are present locally, which is used to avoid unnecessary
data fetches.

Modified during rebase/cherry-pick process.