Checkpoint
Disable propagation of satisfiability to next if Task Finished is received
without a Release Access. If a Release Access is received for an access that is
not topmost (i.e. was shortcutted in the external namespace), then set read and
write satisfiability, topmost bit and received reduction info.

Note: Conflict solved during cherry-pick.