Disable namespace when not allowed
When namespace propagation is disabled, the previous access in the bottom map
gets the noNamespacePropagation status bit set (by setNoNamespacePropagation).
This means that when the access for the new task is linked into the dependency
system, it will be put into the bottom map, but the old access that was in the
bottom map (that has the noNamespacePropagation status bit set) will not point
to it as the next access.  Also the new access gets the topmost and received
reduction info bits set immediately, so that it can be deleted as normal when
its task finishes.