Workaround for deleting remoteTaskInfo holding lock
This is a workaround for the case where the task actually executes and finishes
before leaving remoteTaskCreateAndSubmit (i.e.  before releasing
remoteTaskInfo._lock. When this happens the remoteTaskInfo is actually
destroyed inside the same thread (i.e. as a nested function call inside
AddTask::submitTask or propagateSatisfiability). This raises an assertion in
the debug build (lock destroyed while held by the same thread) and would be
potential memory corruption otherwise.  This is a workaround that stops this
happening, but it does cause a memory leak on the remoteTaskInfos. This
fixes namespacefail1.