Propagate data release if child namespace propagation
Each access of an offloaded task A has a pointer to the appropriate data
release step. If A creates subtasks that access this data, then the pointer to
the data release step is propagated from access to access so that the data
release step is done by the last child task that actually accesses the data.
This propagation is controlled by _propagatesDataReleaseStepToNext. Previously
it was done once the child access becomes complete (plus some other
conditions). This is not correct with remote namespace propagation. Assume that
A in turn offloads two child tasks A1 and A2 to the same node and remote
namespace propagation happens from A1 to A2.  In this case the accesses of A1
never become complete on the node that executes A, so A's data release step is
never passed from A1 to A2 on that node.  The pointer to the release step is
therefore stuck at A1 and it never gets used, so the node that offloaded A
never receives the RELEASE_ACCESS message for that access. This fixes
namespacefail4.c.