Fix workflow "#pragma oss taskwait on" and if0
"#pragma oss taskwait on" is translated by Mercurium into an if0
task. If0 tasks are handled as a special case in the runtime,
which was previously incompatible with the execution workflow.
This patch fixes some incompatibilities with the workflow.
Additionally, in order to correctly handle data fetches in the
cluster version, all if0 tasks, even ready ones, are now executed
non-inline using the workflow, similarly to accelerated if0 tasks.

Note: This commit was modified during rebase.