Extend set personality filter condition.
Add check argv != NULL because with some compilers the constructors are
not receiving argc y argv correctly. This error has been detected in
nord3 when using gcc 4.9, 5.* and 7.*.

This is actually a workaround that must be completed probably with a
configuration option in compile time. Toml config is not possible
because _nanos6_pre_loader is called too early.