Optimize read-only auto accesses.
Use AccessInfo message to hint to the offloader that an auto access is
read-only. Doing it with a hint means that in principle the information can be
sent as soon as the parent task (with the auto dependency) finishes. But that
would need some extra bookkeeping to keep a record on the bottom map whether
the tasks are all read-only, and it would have to be clever if there is another
level of task nesting. For now we send the hint if the taskwait at the end of
the parent task gets read satisfiability before write satisfiability and the
WriteID is the same as the original access.  So unfortunately this hint is on
the critical path; i.e. the sequence on the critical path is 1.  offloader
sends read satisfiability to node j, 2. node j sends AccessInfo with the
ReadOnly hint back to the offloader, 3. offloader sends read satisfiability to
node j+1, ...

This optimization is checked by auto_true13.