Interrupt current taskfor for higher priority ready task
Taskfors generally try to occupy all the cores (on a socket), which can delay
the scheduling of higher priority tasks. This commit modifies the host unsync
scheduler to check if there is a higher priority ready task in the ready queue,
and if so, stop creating collaborators for the current taskfor and schedule the
higher priority task, which itself may be a taskfor. The interrupted taskfors
are added to a list and resumed later when there is no higher priority task.