Allow propagation of inout to in accesses in namespace
We now support remote namespace propagation among all combinations of in, inout
and out except in to inout and in to out. Both of these require the offloader
to enforce the WaR dependency across all nodes that access the data.

The tricky part about supporting inout/out to in is that (1) the inout must
also release the access, because other nodes may concurrently be performing in
accesses (now that multiple namespaces can concurrently propagate the same in
access in their namespace).  Also, (2) the offloader must send read and write
satisfiability for each 'in' access, in case it was not possible to propagate
in the remote namespace. In fact the offloader isn't tracking whether each node
(of possibly many) is propagating the 'in' access in its namespace. So we
propagate read satisfiability in the remote namespace (enabling the access to
become satisfied as soon as possible), while propagating write satisfiability
to the remote tasks via the satisfiability messages.

This means that an inout access followed by an in access (i.e. one with the
"NamespaceNextIsIn" bit set) must both trigger a data release to the offloader
and also propagate to the next in the namespace. Hence the extra condition in
checkDataRelease to allow this to happen.