Allow concurrent reads in multiple namespaces
The idea is to let the namespace predecessor be the ID of the task that wrote
the data. Previously it was always the immediate predecessor. Now the remote
namespace can propagate from task-to-task if either (1) the previous task is
the new task's namespace predecessor or (2) it is a read access whose namespace
predecessor is the same as the new task's namespace predecessor. This only
works now that the OffloadedTaskNum is globally unique. It wouldn't have worked
using the Task pointer on the offloader node, since the same pointer can be
reused and therefore refer to different tasks.

We can still only propagate from an in to in access or from a non-in to non-in
access. Previously this was enforced at the remote end, but the remote end may
see inout -> inout and not be aware that there is an 'in' access in-between. So
the enforcement is now in applyUpdateOperationOnAccess at the offloader side.
Another thing to be careful about is that we must connect read accesses in the
same order as the sequential program, otherwise the later access (in sequential
order) will never receive satisfiability information (as that is passed from
access to next access). We make sure accesses are connected in order in the
remote namespace using the new "namespaceReaderNum" counter.