Check for read satisf. propagation on exit taskwait
This fixes a bug introduced in:

   Only propagate read satisfiability through namespace

The bug occurs in namespacefail11 (roughly 1 in 3 executions).

The bug was introduced in the above commit, which delays the propagation of
read satisfiability while the location in non-local. The intention is to avoid
redundant data fetches that would otherwise happen when multiple tasks access
the same location with an in access, and all would fetch from the original
location. The problem was that when the data location is transferred from a
taskwait to the original access, it may be necessary to propagate read
satisfiability to the next task, but the code was not checking for this.