Extrae workaround: lock around MPI calls if built with Extrae
The issue was reported to tools team by email on 21 Feb 2021.  It seems to be a
race condition in multi-threaded code, and is likely happening now rather than
before because of the dedicated LeaderThread constantly calling MPI_Testsome
while other threads perform sends and receives.  It seems to happen when an
MPI_Request is freed and the same MPI_Request is recycled from a receive to a
send by the MPI library. Then (inside Extrae), getCommDataFromStatus is called
on the MPI_Status corresponding to an MPI_Isend rather than an MPI_Irecv. So
local_source in getCommDataFromStatus is out-of-range and src_world becomes
invalid in ProcessRequest, causing a segfault on the lookup in TracingBitmap in
REAL_TRACE_MPIEVENT_NOHWC.

In Nanos6@cluster, the MPI_Testsome fails when a MessageDataFetch completes.
But checking inside Extrae shows that the same MPI_Request value passed to
MPI_Testsome (for a MessageDataFetch) is recognized as a receive and sent to
getCommDataFromStatus, and that the communicator is INTRA_COMM_DATA_RAW, so the
MPI_Request is handled as if it were a DATA_RAW (likely to be a receive).
Adding the lock (only when built for Extrae) does impact performance but it
does not seem to be excessive: Infoli is 15.2 ms per iteration with Extrae and
14.2 ms without (so 7% slowdown).