Separate DataAccess fields for previous and own namespace
In the previous version, a single member _validNamespace of DataAccess was used
for both the namespace of the previous access and the namespace of the current
access. This saves some space but doesn't handle race conditions between
scheduling (which sets the current access's namespace) and data dependencies
(which set the previous access's namespace). It is better to use separate
fields.