Delete access only after previous namespace info received
This fixes a common segfault in Jacobi, due to race conditions between
propagating the namespace information and propagating read and write
satisfiability.  Since the access could sometimes be deleted before the
namespace information was received, this meant that when the namespace
information was eventually propagated, the task (and its access structures) may
already have been deleted. This caused segmentation faults in the AVL sets in
the deconstructed access structures.