Fix createTopLevelSink without autowait
Top-level sinks are needed for each access of an offloaded task that does not
have a next (which would be namespace propagation). Previously a top-level sink
was created for each access that doesn't have early release in the namespace.
These are the same thing when disable_autowait=false (the default), but not
otherwise.