Fix compatibility with old std versions.
std::unordered_map need to hash the KEY element. In old std versions of
the library it seems like the std::hash function breaks when the key is
an enum. This broke the Generator for schedulers that uses the
nano6_cheduler_type_t as the key.

Using a map may be an issue if the factory generates many kinds of
objects... but that can be solved casting the key in the generator
functions and using unordered_map.