Fix registerLocalAccess that overlaps an existing region
Update registerLocalAccess to not assume that the newly registered access does
not overlap an existing region. This assumption was OK before, but it will
fail for a task with a weakinout on "all memory". When such a task does
an lmalloc (or a dmalloc?) the new region will of course overlap an existing
region that came from the "all memory" dependency.