Add cluster.eager_weak_fetch and support for it being false
The default value of cluster.eager_weak_fetch is true, which is the old
behaviour, meaning that the workflow will eagerly fetch a weak access if it is
read satisfied at the time that the task is scheduled. In practice this is
probably not very common because most weak tasks will be scheduled before the
data is ready (this being the main point of a weak task). Nevertheless we keep
the previous behaviour as the default to avoid introducing an unexpected change
in performance.