Simpler check for data copy fetch and registration
This commit puts in one place, in clusterFetchData, the calculations for
whether a data copy between nodes is needed, and if not, whether the current
(or new) location needs to be registered in the dependency system. Previously
clusterFetchData determined whether a data copy is needed and requiresDataFetch
checked whether the location needed registering. This is a preliminary change
before adding the support for cluster.eager_weak_fetch=false.