Allow tasks with weakinout on all of memory
The bulk of a weakinout on "all memory" consists of very large regions of
memory that is not managed by OmpSs@cluster. The default location for all of
these regions is the directory (this is set in the missing processor in
replaceMatchingInBottomMapLinkAndPropagate, which is called by
linkTaskAccesses). This is a reasonable choice as a way to indicate
uninitialized data with no actual location.

The main idea of this change is to pass these dependencies among nodes with the
location still being the directory. The previous code used to convert to the
physical node when setting up the satisfiability information, and it never
passed the location as being in the directory. But in this case there is no
sensible choice for the physical node, and the previous code failed at exactly
this point. Rather than sending an arbitrary node (node 0?), the cleanest
solution is to just say that the location is the directory, with the meaning
being that the data is uninitialized as far as OmpSs@cluster is concerned.