Check for incorrect pragma that may otherwise hang
This implementation was prone to hanging when a subtask of a task other than
main has an access that is not part of its parent's dependencies. Check for this
case and add an error.

Also add the missing registration of an lmalloc in the dependency system when
cluster.communication = "disabled".