Offloaded in accesses use pseudowrite
Pseudowrite: when an offloaded task has an in access, the true write
satisfiability is not sent to the remote node. This avoids a pointless
ping-ponging of write satisfiability. Instead a "fake" write satisfiability
(pseudowrite) is sent to the remote node. The only downside is that it is no
longer possible to propagate satisfiability from an in access to an inout or
out access through the namespace.