disableReadPropagationToNext only if not weak and not offloaded
This is more complicated than it seems, because Task::isOffloadedTask is false
until the task is offloaded. If we disable propagation to next the old way just
conditional on the task being offloaded, then it can happen that when the
access becomes read satisfied, with a non-local location, the task has not yet
been offloaded; which means that read satisfiability will not be propagated
(disableReadPropagation=1). But when the data arrives the task may have been
scheduled and offloaded, meaning that disableReadPropagation=0 both before and
after updating the location. Hence handleDataAccessStatusEffects will see no
change and assume that read satisfiability has already been propagated. This
caused an occasional hang in unit_tests/nasty/test3_disablereadprop.c