Also call releasePendingAccesses in disposeTask
This fixes nasty/test3_hang1. There was a race condition when two threads
perform concurrent updates to an offloaded task, e.g. one thread "A" does
unregisterTaskDataAccesseses while another thread "B" processes
propagateSatisfiability for a weak access that is not on the bottom map. Thread
"A" releases some of the task's accesses and sends the vector from
processDelayedOperationsSameTask. Then thread "B" releases the non-weak access
and adds it to the (now-emptied) vector. Finally, thread "A"s reaches the
callback in the notification step and it actually disposes the task (thread "B"
could not because the body wasn't complete).  In this case, the release access
from thread "B" never gets sent. Adding an extra call to releasePendingAccesses
in disposeTask ensures that all pending release accesses get sent before the
task is disposed.  In almost all cases, i.e. when this race condition doesn't
happen, the list is empty at this point. We keep the original two calls in
DataAccessRegistration, as they handle early releases of pending accesses when
a subtask of an offloaded task completes.