Remove cluster and host notifications separation.
Simple is better than complex.
Complex is better than complicated.

They both have the same code, so avoid the constructor, duplicate code
for two different classes. If we need another class in the future we can
split them back, but atm don't make it complex.

Construct a simpler lambda for collaborators. We know in advance if the
task is a collaborator, so we can construct a simpler notification step
and condition from the beginning. This ads an extra `if` in the workflow
constructor, but reduces the overhead and simplifies debugging.