Fix dependencies of variables on the stack
Dependencies on the stack were probably broken by "Allow tasks with weakinout
on all of memory". It is fixed by adding the stack as an explicit dependency for
every task.  Otherwise the stack was seen as uninitialized memory so was not copied in
to any subtask. This fixes manual/stackvar[1-3].c and MicroPP.