callbackDecrement in the right places.
* src/cluster/offloading/TaskOffloading.cpp : remoteTaskCleanup is the
hook we use for remote tasks in the namespace. So it must call
NodeNamespace::callbackDecrement to be consistent.
* src/cluster/NodeNamespace.cpp (NodeNamespace::NodeNamespace) : Set
_taskInfo.destroy_args_block. The new function decrements calling
NodeNamespace::callbackDecrement, but any other finalization action must
be there.
* src/cluster/NodeNamespace.hpp (destroyArgsBlock) : New function to use
as destroy_args_block hook.
(notifyShutdown) : simplify a condition (WIP fix it in the right way).
*
src/executors/threads/TaskFinalizationImplementation.hpp (TaskFinalization::disposeTask)
: Remove specific code to call NodeNamespace::callbackDecrement(). These
conditions are now in the hooks as they should.