Group Taskfinish, improve lock and solve memory leak
This commit performs several changes all together. They can't be
splitted into smaller commits because all of them are related each
other.

* src/cluster/messages/MessageReleaseAccess.[c,h]pp : Add a new field to
the releaseAccess message to indicate if we should release the also the
task.

This change goups the TaskFinalization message with the
accessReleaseMessages when possible.
TODO: Update the instrumentation to be more useful about these messages.

* src/cluster/offloading/ClusterTaskContext.hpp: Takes ownership of the
TaskNewMessage. In a future commit we will delete the MessageTaskNew
there and solve the memory leak we have there. (Now MessageTaskNew are
never deleted)

* src/cluster/offloading/TaskOffloading.cpp (remoteTaskCleanup) : now
receives the ClusterContext as input. This will make possible to call
the DataReleaseStep from the ClusterContext's hook. That's the right way
to handle this.

* src/dependencies/linear-regions-fragmented/DataAccessRegistration.cpp
: Call releasePendingAccesses conditionally only when disposeTask is not
going to be called for that task. releasePendingAccesses will be called
unconditionally in the task destructor any way; and will be possible to
add the release information too in that moment.

This also solves the order issue we had before. After the task is
released no more accesses will be added or sent for that task.

* src/executors/workflow/cluster/ExecutionWorkflowCluster.hpp : Set the
release info in the Message release access. Also improve the locking
condition.
When the task sets the releaseTask flag then the ReleaseStep (this) is
also deleted. This may solve the memory leak issue we had before when
the ReleaseStep was not deleted in some conditions.

So this change:
1) Improves taking the release lock.
2) Solved the memory leak.
3) Reduces when possible TaskFinalization messages.