Support commutative accesses
Various changes to support commutative accesses, including for offloaded tasks.
Note: tasks with weakcommutative accesses cannot be offloaded, because they
become ready immediately (once all strong accesses are satisfied), so there is
no way to synchronize commutative subtasks on another node. The main changes
are:

1. Keep the location in CommutativeScoreboard. This is needed because normally
the location is passed from access to access in the original sequential order,
which matches the execution order. Commutative accesses are done out-of-order,
but before they can execute they need the latest location for the commutative
access. This also means that the entry in the CommutativeScoreboard cannot be
reclaimed as soon as there are no waiting participating commutative tasks, as
was done before. That would mean that there would be no way to pass the
location to the next commutative task that arrives. Instead there is a new
method, CommutativeScoreboard::endCommutative, which is called explicitly when
either there is a non-commutative successor or the access is released to the
offloader.

2. Offloaded commutative tasks get pseudoread and pseudowrite satisfiability.
Similarly, receiving pseudoread and pseudowrite satisfiability at the remote
end implies commutative satisfiability. This means that the satisfiability
messages do not need an extra commutative satisfiability bit, and it cuts down
on pointless extra satisfiability messages.

3. Remove the restriction in the ExecutionWorkflow.cpp that explicitly skips
the workflow for commutative accesses.