Support concurrent and weakconcurrent accesses
Various changes to support concurrent and weakconcurrent accesses, including
for offloaded tasks. The main changes are:

1. Add a new field concurrentInitialLocation to the data access, which is the
initial location before a set of concurrent accesses, and it is passed with
concurrent satisfiability. A second location is passed when the task becomes
read satisfied, and this may be different if an earlier task has written to the
region.  Lastly, the task itself (or a subtask) may write to it and change the
location. If either of the new locations is different from the initial location
(they cannot both be different), then we take whichever is different, which
must reflect the location of the actual write, otherwise we keep the initial
location.

2. If a concurrent access of an offloaded task ends with the same location as
the initial location, then we do not send a MessageReleaseAccess. This is an
optimization reflecting the fact that no update has happened (maybe a similar
optimization can be done for non-concurrent accesses?).

3. Concurrent accesses also get pseudowrite and pseudoread in the same way
as commutative accesses.

4. Remove the restriction in the ExecutionWorkflow.cpp that explicitly skips
the workflow for concurrent accesses.

5. Don't eager weak fetch weakconcurrent accesses. The idea is that only part
of a weakconcurrent access is likely to be part of a strong access, so eager
weak fetch is unlikely to be helpful.