Tidy update operation setting of satisfiability
Move the common logic in applyUpdateOperationOnAccess that checks whether to
propagate read, write, concurrent and commutative satisfiability into the
access. If it is a propagate satisfiability into an access that actually takes
satisfiability propagated in the namespace then it is ignored. This code was
already repeated twice effectively, and the same is also needed for concurrent
and commutative satisfiability.