Fix unregistration of local regions with all memory
If a task has an "all memory" region, then local regions (stack or lmalloc)
must not be unregistered by unregisterLocalAccess. They should be left alone,
and handled as normal as a subregion of the "all memory" region. Otherwise
there will be multiple problems, e.g. the task might get disposed of before
satisfiability for that subregion of the "all memory" region arrives, causing a
use-after-free.