Taskwaits don't fetch concurrent or if eager_weak_fetch=false
This is the same condition as for a normal access, from clusterFetchData.
Without this fix, (a) all weakconcurrent tasks that do a taskwait effectively
access the whole region, but only one of them should (this breaks correctness
as it is invalid behaviour), and (b) a taskwait in an "all memory" task will
fetch all data that isn't uninitialized (a possible performance issue). Fix (a)
fixes concurrent9. In order for this to work, the BottomMapEntryContents needs
to keep track not only the original DataAccessType, but also whether or not it
was weak.