Disable namespace propagation from concurrent/commutative
For simplicity, do not propagate in the namespace for concurrent accesses.
Arguably this could be a useful optimization, but namespace propagation only
works for tasks that are consecutive in sequential order. Perhaps something
more powerful that knows about concurrent would be better. Don't do the
optimization until something about its requirements is known.