Fix race condition when constructing an offloaded task
There were (at least) two race conditions where another thread could dispose of
a task before remoteTaskCreateAndSubmit had finished (see comment in this
patch). One of them was partially handled using a workaround,
RemoteTaskInfo._taskBeingConstructed. But it was ugly. Remove this and handle
both race conditions by increasing the task's removalBlockingCount while it is
being constructed. At the end, if decreasing removalBlockingCount brings it to
zero, then everything else is done and we can just dispose of the task.

This fixes namespacenewfail24.