Fix occasional hang caused by passing namespace information
DataAccessRegistration::setNamespaceSelf sets the namespace for the current
access and passes the information to the successor access as a delayed
operation.  Unfortunately this was done by releasing the lock on the access
structures and processing the delayed operations right away. Since
setNamespaceSelf is called by the constructor for the ClusterDataLinkStep,
which is (indirectly) called by the lambda passed to
DataAccessRegistration::processAllDataAccesses, releasing the lock meant that
another thread could fragment the access while the loop is going. The outcome
was that only the data link step for the first part of the fragmented access
got constructed; the rest got skipped and its satisfiability messages didn't
get sent, causing the hang.

Fix this by passing the CPUDependencyData through various workflow functions
and calling a new function to do all the namespace-related delayed operations
once all the data accesses have been processed (and the lock has been
released).  This fixes an occasional hang in nasty_test3_stress and
nasty_test3_fail17.