Fix memory issue in extrae instrumentation.
There were 3 problems:
1) TaskNew messages were leaked from memory.
2) Extrae instrumentation had a set of taskinfo pointers that is
accesses at the end of the execution (in instrument::shutdown). That
pointed required to access to some memory of TaskNew; so this issue
enforced the previous one.
3) As the taskinfo was only associated with the pointer value then many
values in the set contained actually repeated information but stored n
different memry addresses. In the SMP version the taskinfo is a static
vale created by mercurium; but in clusters taskinfo is part of the
message, so they have different addresses. This made iterating over
unneeded repeated taskinfos in 2) over leaked memory from 1).

This commit adds a new data-structure (user_fct_t) in the extrae
Namespace with the information needed to register code in extrae (so the
memory of issue 1) can be released); comparing the values not the
memory (solving issue 3) and storing values of the structure not
pointers to potentially released parts in the code, so the user_fct_t
can be accessed in the shutdown (solving issue 2).